requirements.txt
# Core LangChain packages
langchain>=0.1.0,<0.3.0
langchain-community>=0.0.20,<0.3.0
langchain-openai>=0.0.5,<0.3.0
langchain-neo4j>=0.0.1,<0.3.0

# Database
neo4j>=5.0.0,<6.0.0

# Telegram Bot
python-telegram-bot>=20.0,<21.0

# Environment management
python-dotenv>=1.0.0,<2.0.0

# Retry logic
tenacity>=8.0.0,<9.0.0

# Structured logging
structlog>=23.0.0,<24.0.0

# Testing (development)
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-mock>=3.10.0,<4.0.0

# Code quality (development)
black>=23.0.0,<24.0.0
ruff>=0.1.0,<1.0.0
mypy>=1.5.0,<2.0.0

# Monitoring (optional)
prometheus-client>=0.18.0,<1.0.0
